twitter-localization/twitter-localization-backend/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
neo4j-driver==1.7.1
neobolt==1.7.3
neotime==1.7.4
oauthlib==3.0.1
pymongo==3.7.2
PySocks==1.6.8
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
tweepy==3.7.0
urllib3==1.24.1
Werkzeug==0.14.1
